Enabled switching to tags by clicking
This is a little fragile, but it seems like it works at least for my
configuration. If you do a SUPER+SHIFT+CLICK1 on a window, it will
activate all the tags associated with that window. The main idea here
is to allow going to the tags of a window from view all tags with a
mouse click.